Implement StripePaymentProvider with retry logic and session support

+++ src/main/java/com/adamo/vrspfab/payments/PaymentRequestDto.java
@@ -3,12 +3,11 @@ package com.adamo.vrspfab.payments;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 @Data
 public class PaymentRequestDto {
     private Long reservationId;
-    private BigDecimal amount;
+    private long amount;
     private String currency;
     private String paymentMethodId;
 }
